fix issue with redacting hashes
